+++ client/src/components/settings/Settings.module.css
@@ -7,67 +7,3 @@
   width: 250px;
   padding: 1rem;
 }
-
-.settingsTabContainer {
-  width: 500px;
-}
-
-.input {
-  padding-left: 1rem;
-  width: 100%;
-  height: 40px;
-  border-radius: 20px;
-  margin-bottom: 0.5rem;
-}
-
-.input--light {
-  background-color: var(--chakra-colors-gray-100);
-  box-sizing: border-box;
-  -moz-box-sizing: border-box;
-  -webkit-box-sizing: border-box;
-}
-
-.input--dark {
-  background-color: var(--chakra-colors-gray-900);
-  box-sizing: border-box;
-  -moz-box-sizing: border-box;
-  -webkit-box-sizing: border-box;
-  border: 1px solid var(--chakra-colors-gray-700);
-}
-
-.fileInput {
-  margin-top: 1rem;
-  border-radius: 20px;
-  width: 100%;
-}
-
-.fileInput::file-selector-button {
-  padding: 6px 12px 6px 12px;
-  box-shadow: none;
-  outline: none;
-  border: none;
-  margin-right: 16px;
-  font-weight: 700;
-}
-
-.fileInput::file-selector-button:hover {
-  cursor: pointer;
-}
-
-.fileInput--light {
-  background-color: var(--chakra-colors-gray-200);
-}
-
-.fileInput--dark {
-  background-color: var(--chakra-colors-gray-800);
-}
-
-.fileInput--light::file-selector-button {
-  background-color: var(--chakra-colors-blue-500);
-  color: var(--chakra-colors-white);
-}
-
-.fileInput--dark::file-selector-button {
-  background-color: var(--chakra-colors-blue-200);
-  color: var(--chakra-colors-black);
-}
