+++ client/src/components/settings/adminSettings/index.tsx
@@ -1,0 +1,3 @@
+import GeneralSettings from '../generalSettings/GeneralSettings';
+
+export default GeneralSettings;
