+++ client/src/components/settings/settingsTab/SettingsTab.module.css
@@ -1,0 +1,3 @@
+.settingsTabContainer {
+    width: 500px;
+}
