+++ client/src/components/settings/settingsTab/index.tsx
@@ -1,0 +1,3 @@
+import SettingsTab from './SettingsTab';
+
+export default SettingsTab;
