+++ client/src/components/settings/userSettings/index.tsx
@@ -1,0 +1,3 @@
+import UserSettings from './UserSettings';
+
+export default UserSettings;
